Splitting up files and fixing other errors
