+++ src/BloomBrowserUI/teamCollection/TeamCollectionSettingsPanel.stories.tsx
@@ -1,0 +1,14 @@
+import * as React from "react";
+import { TeamCollectionSettingsPanel } from "./TeamCollectionSettingsPanel";
+
+export default {
+    title: "Team Collection components"
+};
+
+export const _TeamCollectionSettingsPanel = () => (
+    <TeamCollectionSettingsPanel />
+);
+
+_TeamCollectionSettingsPanel.story = {
+    name: "TeamCollectionSettingsPanel"
+};
